+++ src/Admin/Team/Application/UpdateTeamUseCase.php
@@ -50,7 +50,7 @@ final class UpdateTeamUseCase {
             $imageTranslations[] = new TeamImageTranslation($baseLang, null); // Por ahora sin descripción
             $teamImages[] = new TeamImage(
                 $imageData['path'],
-                $imageData['order'] ?? 0,
+                (int) ($imageData['order'] ?? 0),
                 $imageTranslations
             );
         }
